are_identical {assertive.base}	R Documentation

Are the inputs identical?

Description

Checks if the inputs are identical.

Usage

are_identical(x, y, allow_attributes = FALSE,
  .xname = get_name_in_parent(x), .yname = get_name_in_parent(y))

are_identical_legacy(..., l = list())

assert_are_identical(x, y, allow_attributes = FALSE,
  severity = getOption("assertive.severity", "stop"))

assert_all_are_identical_legacy(..., l = list())

assert_any_are_identical_legacy(..., l = list())

Arguments

x
An R object or expression.

y
Another R object or expression.

allow_attributes
If TRUE, The attributes of x and y are allowed to differ.

.xname
Not intended to be used directly.

.yname
Not intended to be used directly.

...
Some R expressions, deprecated.

l
A list of R expressions, deprecated.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

Value

are_identical returns TRUE if x and y are identical. The assert_* function throws an error on failure.

The legacy function are_identical_legacy allows an arbitrary number of inputs and returns a symmetric square logical matrix which is TRUE where pairs of inputs are identical. (The new version of the function is easier to work with, and it is recommended that you switch your code to it.)

See Also

identical, are_same_length

Examples

x <- 1:5
are_identical(c(1, -1), cos(c(0, pi)))
assertive.base::dont_stop(assert_are_identical(c(1, 1), cos(c(0, pi))))

[Package assertive.base version 0.0-7 Index]