dont_stop {assertive.base}	R Documentation

Run code without stopping

Description

Runs code without stopping for warnings or errors.

Usage

dont_stop(expr)

Arguments

expr
Code to execute.

Value

A list containing the results of evaluating each call in expr.

Note

This function is dangerous, since it overrides warnings and errors. Its intended use is for documenting examples of warnings and errors.

See Also

warning and stop for generating warnings and errors respectively; try and conditions for handling them.

Examples

dont_stop({
  warning("a warning")
  x <- 1
  stop("an error")
  y <- sqrt(exp(x + 1))
  assert_is_identical_to_true(y)
  y > 0
})

[Package assertive.base version 0.0-7 Index]