false {assertive.base}	R Documentation

FALSE, with a cause of failure.

Description

Always returns the value FALSE, with a cause attribute.

Usage

false(...)

Arguments

...
Passed to gettextf to create a cause of failure message.

Value

FALSE with the attribute cause, as provided in the input.

See Also

cause and na.

[Package assertive.base version 0.0-7 Index]