get_name_in_parent {assertive.base}	R Documentation

Get the name of a variable in the parent frame

Description

Gets the name of the input in the parent frame.

Usage

get_name_in_parent(x, escape_percent = TRUE)

Arguments

x
Variable to get the name of.

escape_percent
Logical. If TRUE, percent signs are doubled, making the value suitable for use with sprintf (and hence by false and na).

Value

A string giving the name of the input in the parent frame.

Examples

outside <- 1
f <- function(inside, escape_percent) 
{
  get_name_in_parent(inside, escape_percent)
}
f(outside, TRUE) 
f('10%', TRUE) 
f('10%', FALSE)

[Package assertive.base version 0.0-7 Index]