merge.list {assertive.base}	R Documentation

Merge two lists

Description

Merges two lists, taking duplicated elements from the first list.

Usage

## S3 method for class 'list'
merge(x, y, warn_on_dupes = TRUE,
  allow_unnamed_elements = FALSE, ...)

Arguments

x
A list.

y
A list.

warn_on_dupes
TRUE or FALSE. Should a warning be given if both x and y have elements with the same name. See note.

allow_unnamed_elements
TRUE or FALSE. Should unnamed elements be allowed?

...
Ignored.

Value

A list, combining elements from x and y.

Note

In the event of elements that are duplicated between x and y, the versions from x are used.

See Also

merge_dots_with_list, merge

Examples

merge(
  list(foo = 1, bar = 2, baz = 3), 
  list(foo = 4, baz = 5, quux = 6)
)

# If unnamed elements are allowed, they are included at the end
merge(
  list("a", foo = 1, "b", bar = 2, baz = 3, "c"), 
  list(foo = 4, "a", baz = 5, "b", quux = 6, "d"),
  allow_unnamed_elements = TRUE
)

[Package assertive.base version 0.0-7 Index]