merge_dots_with_list {assertive.base}	R Documentation

Merge ellipsis args with a list.

Description

Merges variable length ellipsis arguments to a function with a list argument.

Usage

merge_dots_with_list(..., l = list(), warn_on_dupes = TRUE,
  allow_unnamed_elements = FALSE)

Arguments

...
Some inputs.

l
A list.

warn_on_dupes
TRUE or FALSE. Should a warning be given if both x and y have elements with the same name. See note.

allow_unnamed_elements
TRUE or FALSE. Should unnamed elements be allowed?

Value

A list containing the merged inputs.

Note

If any arguments are present in both the ... and l arguments, the ... version takes preference, and a warning is thrown.

See Also

merge.list, merge

Examples

merge_dots_with_list(
  foo = 1, 
  bar = 2, 
  baz = 3, 
  l = list(foo = 4, baz = 5, quux = 6)
)

[Package assertive.base version 0.0-7 Index]