print.scalar_with_cause {assertive.base}	R Documentation

Print methods for objects with a cause attribute

Description

Prints objects of class scalar_with_cause and vector_with_cause.

Usage

## S3 method for class 'scalar_with_cause'
print(x, ...)

## S3 method for class 'vector_with_cause'
print(x, na_ignore = FALSE, n_to_show = 10, ...)

Arguments

x
an object of class scalar_with_cause or vector_with_cause.

...
Currently unused.

na_ignore
A logical value. If FALSE, NA values are printed; otherwise they do not. Like na.rm in many stats package functions, except that the position of the failing values does not change.

n_to_show
A natural number. The maximum number of failures to show.

[Package assertive.base version 0.0-7 Index]