safe_deparse {assertive.base}	R Documentation

Safe version of deparse

Description

A version of deparse that is guaranteed to always return a single string.

Usage

safe_deparse(expr, ...)

Arguments

expr
Any R expression.

...
Passed to deparse.

Value

A character vector or length one.

Note

By default the RStudio IDE truncates output in the console at 1000 characters. Consequently, if you use safe_deparse on large or complex objects, you won't see the full value. You can change the setting using Tools -> "Global Options..." -> Code -> Display -> Console -> "Limit length of lines displayed in console to:".

Examples

# safe_deparse only differs from deparse when the deparse string is longer
# than width.cutoff
deparse(CO2, width.cutoff = 500L) # has length 6
safe_deparse(CO2)                 # has length 1

[Package assertive.base version 0.0-7 Index]