use_first {assertive.base}	R Documentation

Only use the first element of a vector

Description

If the input is not scalar, then only the first element is returned, with a warning.

Usage

use_first(x, indexer = c("[[", "["), .xname = get_name_in_parent(x))

Arguments

x
Input that should be scalar.

indexer
Either double indexing, "[[" (the default) or single indexing "[".

.xname
Not intended to be used directly.

Value

If x is scalar, it is returned unchanged, otherwise only the first element is returned, with a warning.

Examples

dont_stop(use_first(1:5))

[Package assertive.base version 0.0-7 Index]